requirements.txt
streamlit
requests
openpyxl
google-generativeai
